+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Fixed signature validation for gossip messages when payload contains more than one empty string
 ### Changed
 - Joining height is defined separately for download and for set
+- Resolving checkpoints in queue with max concurrent limit
 
 ## [v2.24.10] 2021-06-29
 ### Changed